+++ lib/WorseReflection/Core/Type/IntNegative.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Phpactor\WorseReflection\Core\Type;
+
+class IntNegative extends IntType
+{
+    public function __toString(): string
+    {
+        return 'negative-int';
+    }
+    public function toPhpString(): string
+    {
+        return 'int';
+    }
+}